refactor: shifting over to the "google-generative-ai-rs" crate. Implemented the Vertex AI provider.
