+++ Cargo.toml
@@ -20,8 +20,8 @@ path = "src/main.rs"
 [dependencies]
 openai-api-rs = "2.1.6"
 # openai-api-rs = { git = "https://github.com/cosmonaut-nz/openai-api-rs.git"}
-google-generative-ai-rs = "0.1.2" 
-# google-generative-ai-rs = { git = "https://github.com/avastmick/google-generative-ai-rs.git", version = "0.1.2" }
+google-generative-ai-rs = "0.1.3" 
+# google-generative-ai-rs = { git = "https://github.com/avastmick/google-generative-ai-rs.git", version = "0.1.3" }
 thiserror = "1.0.50"
 log = "0.4.20"
 env_logger = "0.10.1"
